feat: IFS-2400 move IsyOAuth2Authentifizierungsmanager to core package with @NonNullApi
