--- isy-security/src/main/java/de/bund/bva/isyfact/security/authentication/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Authentication related Classes and Implementations.
- */
-@NonNullApi
-package de.bund.bva.isyfact.security.authentication;
-
-import org.springframework.lang.NonNullApi;
